major refactor
- Remove progressbar module
- Rewrite archive.extract_zip/extract_tar/extract_tree to use progress reporter
- Remove archive.Archive class
- Rewrite cmd_export.inspect to work with readers
- Rewrite migration tests to work without Archive
- move dbexport.zip contents to common.zip_folder (and deprecate)
- move .dbexport.utils.ExportFileFormat to common.config
- refacor django importer to work with progress reporter
- move readers and writers to seperate module
- added files to mypy

still todo: sqlalchemy import and refactor migration
