+++ aiida/tools/importexport/__init__.py
@@ -15,9 +15,9 @@ explore tree: https://github.com/aiidateam/aiida-core/tree/eebef392c81e8b130834a
 Functionality: <tree>/aiida/orm/importexport.py
 Tests: <tree>/aiida/backends/tests/test_export_and_import.py
 """
-
+from .archive import *
 from .dbexport import *
 from .dbimport import *
 from .common import *
 
-__all__ = (dbexport.__all__ + dbimport.__all__ + common.__all__)
+__all__ = (archive.__all__ + dbexport.__all__ + dbimport.__all__ + common.__all__)
