+++ aiida/tools/importexport/archive/__init__.py
@@ -7,3 +7,12 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
+# pylint: disable=wildcard-import,undefined-variable
+# type: ignore
+"""Readers and writers for archive formats, that work independently of a connection to an AiiDA profile."""
+
+from .common import *
+from .readers import *
+from .writers import *
+
+__all__ = (readers.__all__ + writers.__all__ + common.__all__)
