+++ aiida/tools/importexport/archive/readers.py
@@ -1,0 +1,505 @@
+# -*- coding: utf-8 -*-
+###########################################################################
+# Copyright (c), The AiiDA team. All rights reserved.                     #
+# This file is part of the AiiDA code.                                    #
+#                                                                         #
+# The code is hosted on GitHub at https://github.com/aiidateam/aiida-core #
+# For further information on the license, see the LICENSE.txt file        #
+# For further information please visit http://www.aiida.net               #
+###########################################################################
+"""Archive reader classes."""
+from abc import ABC, abstractmethod
+import json
+import os
+from pathlib import Path
+from typing import Any, cast, Dict, Iterable, Iterator, List, Optional, Set, Tuple, Type
+import zipfile
+import tarfile
+
+from distutils.version import StrictVersion
+
+from aiida.common.log import AIIDA_LOGGER
+from aiida.common.exceptions import InvalidOperation
+from aiida.common.folders import Folder, SandboxFolder
+from aiida.common.progress_reporter import get_progress_reporter
+from aiida.tools.importexport.common.config import EXPORT_VERSION, ExportFileFormat, NODES_EXPORT_SUBFOLDER
+from aiida.tools.importexport.common.exceptions import (CorruptArchive, IncompatibleArchiveVersionError)
+from aiida.tools.importexport.archive.common import ArchiveMetadata
+from aiida.tools.importexport.common.utils import export_shard_uuid
+
+__all__ = (
+    'ArchiveReaderAbstract',
+    'ARCHIVE_READER_LOGGER',
+    'ReaderJsonFolder',
+    'ReaderJsonTar',
+    'ReaderJsonZip',
+    'get_reader',
+)
+
+ARCHIVE_READER_LOGGER = AIIDA_LOGGER.getChild('archive.reader')
+
+
+def get_reader(file_format: str) -> Type['ArchiveReaderAbstract']:
+    """Return the available writer classes."""
+    readers = {
+        ExportFileFormat.ZIP: ReaderJsonZip,
+        ExportFileFormat.TAR_GZIPPED: ReaderJsonTar,
+        'folder': ReaderJsonFolder,
+    }
+
+    if file_format not in readers:
+        raise ValueError(
+            f'Can only read in the formats: {tuple(readers.keys())}, please specify one for "file_format".'
+        )
+
+    return cast(Type[ArchiveReaderAbstract], readers[file_format])
+
+
+class ArchiveReaderAbstract(ABC):
+    """An abstract interface for AiiDA archive readers.
+
+    An ``ArchiveReader`` implementation is intended to be used with a context::
+
+        with ArchiveReader(filename) as reader:
+            reader.entity_count('Node')
+
+    """
+
+    def __init__(self, filename: str, **kwargs: Any):
+        """An archive writer
+
+        :param filename: the filename (possibly including the absolute path)
+            of the file on which to import.
+
+        """
+        # pylint: disable=unused-argument
+        self._filename = filename
+        self._in_context = False
+
+    @property
+    def filename(self) -> str:
+        """Return the filename to write to."""
+        return self._filename
+
+    @property
+    @abstractmethod
+    def file_format_verbose(self) -> str:
+        """The file format name."""
+
+    @property
+    @abstractmethod
+    def compatible_export_version(self) -> str:
+        """Return the export version that this reader is compatible with."""
+
+    def __enter__(self) -> 'ArchiveReaderAbstract':
+        self._in_context = True
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self._in_context = False
+
+    def assert_within_context(self):
+        """Assert that the method is called within a context.
+
+        :raises: `~aiida.common.exceptions.InvalidOperation`: if not called within a context
+        """
+        if not self._in_context:
+            raise InvalidOperation('the ArchiveReader method should be used within a context')
+
+    @property
+    @abstractmethod
+    def export_version(self) -> str:
+        """Return the export version.
+
+        :raises `~aiida.tools.importexport.common.exceptions.CorruptArchive`: If the version cannot be retrieved.
+        """
+        # this should be able to be returned independent of any metadata validation
+
+    def check_version(self):
+        """Check the version compatibility of the archive.
+
+        :raises: `~aiida.tools.importexport.common.exceptions.IncompatibleArchiveVersionError`:
+            If the version is not compatible
+
+        """
+        file_version = StrictVersion(self.export_version)
+        expected_version = StrictVersion(self.compatible_export_version)
+
+        try:
+            if file_version != expected_version:
+                msg = f'Archive file version is {file_version}, can read only version {expected_version}'
+                if file_version < expected_version:
+                    msg += "\nUse 'verdi export migrate' to update this archive file."
+                else:
+                    msg += '\nUpdate your AiiDA version in order to import this file.'
+
+                raise IncompatibleArchiveVersionError(msg)
+        except AttributeError:
+            msg = (
+                f'Archive file version is {self.export_version}, '
+                f'can read only version {self.compatible_export_version}'
+            )
+            raise IncompatibleArchiveVersionError(msg)
+
+    @property
+    @abstractmethod
+    def metadata(self) -> ArchiveMetadata:
+        """Return the full (validated) export metadata."""
+
+    @property
+    def entity_names(self) -> List[str]:
+        """Return list of all entity names."""
+        return list(self.metadata.all_fields_info.keys())
+
+    @abstractmethod
+    def entity_count(self, name: str) -> int:
+        """Return the count of an entity or None if not contained in the archive."""
+
+    @property
+    @abstractmethod
+    def link_count(self) -> int:
+        """Return the count of links."""
+
+    @abstractmethod
+    def iter_entity_fields(self,
+                           name: str,
+                           fields: Optional[Tuple[str, ...]] = None) -> Iterator[Tuple[int, Dict[str, Any]]]:
+        """Iterate over entities and yield their pk and database fields."""
+
+    @abstractmethod
+    def iter_node_uuids(self) -> Iterator[str]:
+        """Iterate node UUIDs."""
+
+    @abstractmethod
+    def iter_group_uuids(self) -> Iterator[Tuple[str, Set[str]]]:
+        """Iterate group UUIDs and the a set of the node UUIDs they contain."""
+
+    @abstractmethod
+    def iter_link_data(self) -> Iterator[dict]:
+        """Iterate links: {'input': <UUID>, 'output': <UUID>, 'label': <LABEL>, 'type': <TYPE>}"""
+
+    @abstractmethod
+    def iter_node_repos(self,
+                        uuids: Iterable[str],
+                        progress: bool = True,
+                        description='Iterating node repos') -> Iterator[Folder]:
+        """Yield temporary folders containing the contents of the repository for each node.
+
+        :param uuids: The node UUIDs to yield
+        :param progress: report progress
+        :param description: description for progress report
+
+        :raises `~aiida.tools.importexport.common.exceptions.CorruptArchive`: If the repository does not exist.
+        """
+
+    def node_repository(self, uuid: str) -> Folder:
+        """Return a temporary folder containing the contents of the repository for a single node.
+
+        :raises `~aiida.tools.importexport.common.exceptions.CorruptArchive`: If the repository does not exist.
+        """
+        return next(self.iter_node_repos([uuid], progress=False))
+
+
+class ReaderJsonBase(ArchiveReaderAbstract):
+    """A reader base for the JSON compressed formats."""
+
+    FILENAME_DATA = 'data.json'
+    FILENAME_METADATA = 'metadata.json'
+    REPO_FOLDER = NODES_EXPORT_SUBFOLDER
+
+    def __init__(self, filename: str, sandbox_in_repo: bool = False, **kwargs: Any):
+        """A writer for zipped archives.
+
+        :param filename: the filename (possibly including the absolute path)
+            of the file on which to export.
+        :param sandbox_in_repo: Create the temporary uncompressed folder within the aiida repository
+
+        """
+        super().__init__(filename, **kwargs)
+        self._metadata = None
+        self._data = None
+        # a temporary folder used to extract the file tree
+        self._sandbox: Optional[SandboxFolder] = None
+        self._sandbox_in_repo = sandbox_in_repo
+
+    @property
+    def file_format_verbose(self) -> str:
+        raise NotImplementedError()
+
+    @property
+    def compatible_export_version(self) -> str:
+        return EXPORT_VERSION
+
+    def __enter__(self):
+        super().__enter__()
+        self._sandbox = SandboxFolder(self._sandbox_in_repo)
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self._sandbox.erase()
+        self._sandbox: Optional[SandboxFolder] = None
+        self._metadata = None
+        self._data = None
+        super().__exit__(exc_type, exc_val, exc_tb)
+
+    def _get_metadata(self):
+        """Retrieve the metadata JSON."""
+        raise NotImplementedError()
+
+    def _get_data(self):
+        """Retrieve the data JSON."""
+        raise NotImplementedError()
+
+    def _extract(self, *, path_prefix: str, progress: bool):
+        """Extract repository data to a temporary folder.
+
+        :param path_prefix: Only extract paths starting with this prefix.
+        :param progress: Whether to report progress of the extraction
+
+        :raises TypeError: if parameter types are not respected
+        """
+        raise NotImplementedError()
+
+    @property
+    def export_version(self) -> str:
+        metadata = self._get_metadata()
+        if 'export_version' not in metadata:
+            raise CorruptArchive('export_version missing from metadata.json')
+        return metadata['export_version']
+
+    @property
+    def metadata(self) -> ArchiveMetadata:
+        metadata = self._get_metadata()
+        export_parameters = metadata.get('export_parameters', {})
+        output = {
+            'export_version': metadata['export_version'],
+            'aiida_version': metadata['aiida_version'],
+            'all_fields_info': metadata['all_fields_info'],
+            'unique_identifiers': metadata['unique_identifiers'],
+            'graph_traversal_rules': export_parameters.get('graph_traversal_rules', None),
+            'entities_starting_set': export_parameters.get('entities_starting_set', None),
+            'include_comments': export_parameters.get('include_comments', None),
+            'include_logs': export_parameters.get('include_logs', None),
+            'conversion_info': metadata.get('conversion_info', [])
+        }
+        try:
+            return ArchiveMetadata(**output)
+        except TypeError as error:
+            raise CorruptArchive(f'Metadata invalid: {error}')
+
+    def entity_count(self, name: str) -> int:
+        data = self._get_data().get('export_data', {}).get(name, {})
+        return len(data)
+
+    @property
+    def link_count(self) -> int:
+        return len(self._get_data()['links_uuid'])
+
+    def iter_entity_fields(self,
+                           name: str,
+                           fields: Optional[Tuple[str, ...]] = None) -> Iterator[Tuple[int, Dict[str, Any]]]:
+        if name not in self.entity_names:
+            raise ValueError(f'Unknown entity name: {name}')
+        data = self._get_data()['export_data'].get(name, {})
+        if name == 'Node':
+            # here we merge in the attributes and extras before yielding
+            attributes = self._get_data().get('node_attributes', {})
+            extras = self._get_data().get('node_extras', {})
+            for pk, all_fields in data.items():
+                if pk not in attributes:
+                    raise CorruptArchive(f'Unable to find attributes info for Node with Pk={pk}')
+                if pk not in extras:
+                    raise CorruptArchive(f'Unable to find extra info for Node with Pk={pk}')
+                all_fields = {**all_fields, **{'attributes': attributes[pk], 'extras': extras[pk]}}
+                if fields is not None:
+                    all_fields = {k: v for k, v in all_fields.items() if k in fields}
+                yield int(pk), all_fields
+        else:
+            for pk, all_fields in data.items():
+                if fields is not None:
+                    all_fields = {k: v for k, v in all_fields.items() if k in fields}
+                yield int(pk), all_fields
+
+    def iter_node_uuids(self) -> Iterator[str]:
+        for _, fields in self.iter_entity_fields('Node', fields=('uuid',)):
+            yield fields['uuid']
+
+    def iter_group_uuids(self) -> Iterator[Tuple[str, Set[str]]]:
+        for key, values in self._get_data()['groups_uuid'].items():
+            yield key, set(values)
+
+    def iter_link_data(self) -> Iterator[dict]:
+        for value in self._get_data()['links_uuid']:
+            yield value
+
+    def iter_node_repos(self,
+                        uuids: Iterable[str],
+                        progress: bool = True,
+                        description='Iterating node repositories') -> Iterator[Folder]:
+        path_prefixes = [os.path.join(self.REPO_FOLDER, export_shard_uuid(uuid)) for uuid in uuids]
+
+        if not path_prefixes:
+            return
+        self.assert_within_context()
+        assert self._sandbox is not None  # required by mypy
+
+        # unarchive the common folder if it does not exist
+        common_prefix = os.path.commonprefix(path_prefixes)
+        if not self._sandbox.get_subfolder(common_prefix).exists():
+            self._extract(path_prefix=common_prefix, progress=progress)
+
+        if progress:
+            with get_progress_reporter()(total=len(path_prefixes), desc=description) as report:
+                for uuid, path_prefix in zip(uuids, path_prefixes):
+                    report.update()
+                    subfolder = self._sandbox.get_subfolder(path_prefix)
+                    if not subfolder.exists():
+                        raise CorruptArchive(
+                            f'Unable to find the repository folder for Node with UUID={uuid} in the archive'
+                        )
+                    yield subfolder
+        else:
+            for uuid, path_prefix in zip(uuids, path_prefixes):
+                subfolder = self._sandbox.get_subfolder(path_prefix)
+                if not subfolder.exists():
+                    raise CorruptArchive(
+                        f'Unable to find the repository folder for Node with UUID={uuid} in the exported file'
+                    )
+                yield subfolder
+
+
+class ReaderJsonZip(ReaderJsonBase):
+    """A reader for a JSON zip compressed format."""
+
+    @property
+    def file_format_verbose(self) -> str:
+        return 'JSON (zip compressed)'
+
+    def _get_metadata(self):
+        if self._metadata is None:
+            try:
+                self._metadata = json.loads(
+                    zipfile.ZipFile(self.filename, 'r', allowZip64=True).read(self.FILENAME_METADATA).decode('utf8')
+                )
+            except zipfile.BadZipfile:
+                raise ValueError('The input file format is not valid (not a zip file)')
+            except KeyError:
+                raise CorruptArchive(f'required file `{self.FILENAME_METADATA}` is not included')
+        return self._metadata
+
+    def _get_data(self):
+        if self._data is None:
+            try:
+                self._data = json.loads(
+                    zipfile.ZipFile(self.filename, 'r', allowZip64=True).read(self.FILENAME_DATA).decode('utf8')
+                )
+            except zipfile.BadZipfile:
+                raise ValueError('The input file format is not valid (not a zip file)')
+            except KeyError:
+                raise CorruptArchive(f'required file {self.FILENAME_DATA} is not included')
+        return self._data
+
+    def _extract(self, *, path_prefix: str, progress: bool):
+        self.assert_within_context()
+        assert self._sandbox is not None  # required by mypy
+        try:
+            with zipfile.ZipFile(self.filename, 'r', allowZip64=True) as handle:
+                members = [m for m in handle.namelist() if m.startswith(path_prefix)]
+                if progress:
+                    with get_progress_reporter()(total=len(members), desc='Extracting repository files') as report:
+                        for membername in members:
+                            report.update()
+                            handle.extract(path=self._sandbox.abspath, member=membername)
+                else:
+                    for membername in members:
+                        handle.extract(path=self._sandbox.abspath, member=membername)
+        except zipfile.BadZipfile:
+            raise TypeError('The input file format is not valid (not a zip file)')
+
+
+class ReaderJsonTar(ReaderJsonBase):
+    """A reader for a JSON tar compressed format."""
+
+    @property
+    def file_format_verbose(self) -> str:
+        return 'JSON (tar.gz compressed)'
+
+    def _get_metadata(self):
+        if self._metadata is None:
+            try:
+                with tarfile.open(self.filename, 'r:*', format=tarfile.PAX_FORMAT) as handle:
+                    self._metadata = json.loads(handle.extractfile(self.FILENAME_METADATA).read())
+            except tarfile.ReadError:
+                raise ValueError('The input file format is not valid (not a tar file)')
+            except (KeyError, AttributeError):
+                raise CorruptArchive(f'required file `{self.FILENAME_METADATA}` is not included')
+        return self._metadata
+
+    def _get_data(self):
+        if self._data is None:
+            try:
+                with tarfile.open(self.filename, 'r:*', format=tarfile.PAX_FORMAT) as handle:
+                    self._data = json.loads(handle.extractfile(self.FILENAME_DATA).read())
+            except tarfile.ReadError:
+                raise ValueError('The input file format is not valid (not a tar file)')
+            except (KeyError, AttributeError):
+                raise CorruptArchive(f'required file `{self.FILENAME_DATA}` is not included')
+        return self._data
+
+    def _extract(self, *, path_prefix: str, progress: bool):
+        self.assert_within_context()
+        assert self._sandbox is not None  # required by mypy
+        try:
+            with tarfile.open(self.filename, 'r:*', format=tarfile.PAX_FORMAT) as handle:
+                members = [m for m in handle.getmembers() if m.name.startswith(path_prefix)]
+                if progress:
+                    with get_progress_reporter()(total=len(members), desc='Extracting repository files') as report:
+                        for member in members:
+                            report.update()
+                            if member.isdev():
+                                # safety: skip if character device, block device or FIFO
+                                msg = f'WARNING, device found inside the import file: {member.name}'
+                                ARCHIVE_READER_LOGGER.warning(msg)
+                            if member.issym() or member.islnk():
+                                # safety: although dereference=True set in export, so this should not occur
+                                msg = f'WARNING, symlink found inside the import file: {member.name}'
+                                ARCHIVE_READER_LOGGER.warning(msg)
+                                continue
+                            handle.extract(path=self._sandbox.abspath, member=member.name)
+                else:
+                    for membername in members:
+                        handle.extract(path=self._sandbox.abspath, member=membername)
+        except zipfile.BadZipfile:
+            raise TypeError('The input file format is not valid (not a zip file)')
+
+
+class ReaderJsonFolder(ReaderJsonBase):
+    """A reader for a JSON plain folder format."""
+
+    @property
+    def file_format_verbose(self) -> str:
+        return 'JSON (folder)'
+
+    def _get_metadata(self):
+        if self._metadata is None:
+            path = Path(self.filename) / self.FILENAME_METADATA
+            if not path.exists():
+                raise CorruptArchive(f'required file `{self.FILENAME_METADATA}` is not included')
+            self._metadata = json.loads(path.read_text(encoding='utf8'))
+        return self._metadata
+
+    def _get_data(self):
+        if self._data is None:
+            path = Path(self.filename) / self.FILENAME_DATA
+            if not path.exists():
+                raise CorruptArchive(f'required file `{self.FILENAME_DATA}` is not included')
+            self._data = json.loads(path.read_text(encoding='utf8'))
+        return self._data
+
+    def _extract(self, *, path_prefix: str, progress: bool):
+        # pylint: disable=unused-argument
+        self.assert_within_context()
+        assert self._sandbox is not None  # required by mypy
+        # By copying the contents of the source directory, we do not risk to modify the source files accidentally
+        # Use path_prefix? or is this quick enough to not worry
+        self._sandbox.replace_with_folder(self.filename, overwrite=True)
