+++ aiida/tools/importexport/dbexport/utils.py
@@ -9,7 +9,6 @@
 ###########################################################################
 """ Utility functions for export of AiiDA entities """
 # pylint: disable=too-many-locals,too-many-branches,too-many-nested-blocks
-from enum import Enum
 import warnings
 
 from aiida.orm import QueryBuilder, ProcessNode
@@ -22,12 +21,6 @@ from aiida.tools.importexport.common.config import (
 )
 
 EXPORT_LOGGER = AIIDA_LOGGER.getChild('export')
-
-
-class ExportFileFormat(str, Enum):
-    """Archive file formats"""
-    ZIP = 'zip'
-    TAR_GZIPPED = 'tar.gz'
 
 
 def fill_in_query(partial_query, originating_entity_str, current_entity_str, tag_suffixes=None, entity_separator='_'):
